+++ schemaorg/templates/google/__init__.py
@@ -42,6 +42,12 @@ def make_person(name, description, url="", telephone="", email=""):
     person.add_property('contactPoint', contactPoint)
     return person
 
+def make_template_base(schema, template, pretty_print):
+    template = get_template(template)
+    metadata = schema.dump_json(pretty_print)
+    template = template.replace("{{ SCHEMAORG_JSON }}", metadata)
+    return template    
+
 
 def make_dataset(schema,
                  output_file=None,
@@ -53,9 +59,26 @@ def make_dataset(schema,
        to change these variables. If an output file is provided, we write the
        template to the file. Otherwise, we just return it.
     '''
-    template = get_template(template)
-    metadata = schema.dump_json(pretty_print)
-    template = template.replace("{{ SCHEMAORG_JSON }}", metadata)
+    template = make_template_base(schema, template, pretty_print)
+    if output_file is not None:
+        write_file(output_file, template)
+    return template
+
+
+def make_table(schema,
+               rows,                 
+               output_file=None,
+               pretty_print=True,
+               title=None,
+               template="google/visual-table.html"):
+
+    '''write a data catalog, meaning rows of data. You should include, as rows,
+       a list of rows as you want them (e.g., each with <tr><td>...</tr></td>.
+       If an output file is provided, we write to file.
+    '''
+    template = make_template_base(schema, template, pretty_print)
+    template = template.replace("{{ SCHEMAORG_TITLE }}", title or "Schema.org Data Catalog")
+    template = template.replace('{{ SCHEMAORG_TABLE }}', '\n'.join(rows))
     if output_file is not None:
         write_file(output_file, template)
     return template